fix: default completionAutonomyLevel to 3 and remove legacy WorkflowEditor
- SpaceWorkflowRepository.createWorkflow now defaults completionAutonomyLevel to
  3 (supervised) when not provided instead of throwing, matching the DB column
  default. This fixes the "completionAutonomyLevel is required" error when
  creating workflows from the visual editor.

- Extracted WorkflowTemplate types and utility functions (filterAgents,
  buildTemplateNodes, getAvailableTemplates, workflowToTemplate) into a new
  workflow-templates.ts file. Updated VisualWorkflowEditor.tsx import path.

- Deleted the legacy WorkflowEditor.tsx component and its test file (the
  component was unused; SpaceConfigurePage only imports VisualWorkflowEditor).

- Added workflow-templates.test.ts with 17 unit tests covering the extracted
  utilities.
