+++ packages/web/src/components/space/visual-editor/VisualWorkflowEditor.tsx
@@ -24,8 +24,8 @@ import { useState, useMemo, useCallback, useRef } from 'preact/hooks';
 import type { SpaceWorkflow, WorkflowNode, WorkflowChannel, Gate } from '@neokai/shared';
 import { generateUUID, TASK_AGENT_NODE_ID, isChannelCyclic } from '@neokai/shared';
 import { spaceStore } from '../../../lib/space-store';
-import { filterAgents, buildTemplateNodes, getAvailableTemplates } from '../WorkflowEditor';
-import type { WorkflowTemplate } from '../WorkflowEditor';
+import { filterAgents, buildTemplateNodes, getAvailableTemplates } from '../workflow-templates';
+import type { WorkflowTemplate } from '../workflow-templates';
 import { ConfirmModal } from '../../ui/ConfirmModal';
 import type { NodeDraft, AgentTaskState } from '../WorkflowNodeCard';
 import type { ViewportState, Point, VisualTransition, WorkflowConditionType } from './types';
